MA-2183 Make SUCCESS_DISPLAY_MS Int+Long widening explicit
Per review: CROSSFADE_MS (Int) + SUCCESS_BUFFER_MS (Long) compiles via
widening but is a micro-surprise for anyone later doing arithmetic on
CROSSFADE_MS. Add an explicit .toLong() at the arithmetic site.

+++ Android/app/src/main/java/com/dmdbrands/gurus/weight/features/common/components/SetupLoader.kt
@@ -51,7 +51,7 @@ object SetupLoaderTimings {
    * navigating away. Covers [CROSSFADE_MS] plus [SUCCESS_BUFFER_MS] for slower
    * devices so the user reliably perceives the checkmark.
    */
-  const val SUCCESS_DISPLAY_MS: Long = CROSSFADE_MS + SUCCESS_BUFFER_MS
+  const val SUCCESS_DISPLAY_MS: Long = CROSSFADE_MS.toLong() + SUCCESS_BUFFER_MS
 }
 
 /** Number of dots rendered by [SetupLoader]. */
